Edited part about what factors are important
I edited the part where I outlined the relevant factors for what would
encourage policy.  One mistake I think I made is that I talked a lot
about coalitions. I need to generalize back out and stop talking about
coalitions. What a waste of time. Just say that certain policy making
tools are closed off based on the configurations of the day. We can
estimate this with some sort of regression and then drawing predictions.
Perhaps coalitions are a part of the story but that is too difficult to
do right now. Just stick to "certain configurations" of these interests
produce different policies.